Need to be fixed: adding manager

+++ SQLI_Gestion_RH_Web/src/main/java/ma/ensao/youmna/service/TechnologieService.java
@@ -14,6 +14,6 @@ public interface TechnologieService {
 	
 	List<String> technologies();
 	
-	Map<String, Integer> getCountTechnologie();
+	Map<String, Long> getCountTechnologie();
 
 }
